requirements.txt
xmltodict
dnspython
shodan>=1.31.0
argparse
selenium>=4.15.1
Pillow>=10.0.1
flask>=3.0.0
requests>=2.31.0
gunicorn
beautifulsoup4>=4.12.2
pyOpenSSL>=23.3.0
python-dotenv>=1.0.0
vt-py
aiohttp>=3.9.0 # not directly required, pinned by Snyk to avoid a vulnerability